backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
drf-spectacular==0.28.0
